moved secrets methods to Configuration class; made environment variable separator configurable;
